+++ tests/test_utils.py
@@ -1,6 +1,7 @@
 import pytest
 
-from bedoner.utils import split_keepsep, zero_pad
+from bedoner.utils import split_keepsep, zero_pad, get_doc_char_span
+from spacy.tokens import Doc
 
 
 def test_zero_pad():
@@ -27,3 +28,13 @@ def test_zero_pad():
 )
 def test_split_keepsep(text, sep, expected):
     assert split_keepsep(text, sep) == expected
+
+
+@pytest.mark.parametrize(
+    "tokens,i,j,expected",
+    [(["Foo", "bar", "baz"], 0, 2, "Fo"), (["Foo", "bar", "baz"], 0, 5, "Foo b")],
+)
+def test_get_doc_char_span(vocab, tokens, i, j, expected):
+    doc = Doc(vocab, tokens)
+    span = get_doc_char_span(doc, i, j)
+    assert span.text == expected